third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/gru.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] gru float32 tensors steps=2 with options.bias, options.recurrentBias, options.direction='backward', options.activations=['relu', 'relu'] and options.returnSequence=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float32 actual -0.25 should be close enough to expected -0.24974998831748962 by ULP distance: expected a number less than or equal to 6n but got 16778n
[FAIL] gru float32 tensors steps=2 with options.bias, options.recurrentBias, options.direction='both' and options.returnSequence=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float32 actual -0.375 should be close enough to expected -0.28068751096725464 by ULP distance: expected a number less than or equal to 6n but got 3164602n
[FAIL] gru float32 tensors steps=2 with all options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float32 actual -0.25 should be close enough to expected -0.24974998831748962 by ULP distance: expected a number less than or equal to 6n but got 16778n
[FAIL] gru float16 tensors steps=1 with options.bias, options.recurrentBias and options.activations=['relu', 'relu']
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.25 by ULP distance: expected a number less than or equal to 6 but got 46080
[FAIL] gru float16 tensors steps=1 with options.bias, options.recurrentBias and options.activations=['relu', 'relu'] and resetAfter=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.375 by ULP distance: expected a number less than or equal to 6 but got 46592
[FAIL] gru float16 tensors steps=1 with options.bias, options.recurrentBias, options.activations=['relu', 'relu'] and explicit options.direction='forward'
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.25 by ULP distance: expected a number less than or equal to 6 but got 46080
[FAIL] gru float16 tensors steps=1 with options.bias, options.recurrentBias, options.activations=['relu', 'relu'] and explicit options.layout='zrn'
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.25 by ULP distance: expected a number less than or equal to 6 but got 46080
[FAIL] gru float16 tensors steps=1 with options.bias, options.recurrentBias, options.activations=['relu', 'relu'] and options.layout='rzn'
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.25 by ULP distance: expected a number less than or equal to 6 but got 46080
[FAIL] gru float16 tensors steps=1 with options.bias, options.recurrentBias, options.activations=['relu', 'relu'] and options.initialHiddenState
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.25 by ULP distance: expected a number less than or equal to 6 but got 46080
[FAIL] gru float16 tensors steps=1 all options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual 0 should be close enough to expected -0.25 by ULP distance: expected a number less than or equal to 6 but got 46080
[FAIL] gru float16 tensors steps=2 with options.bias, options.recurrentBias, options.direction='backward', options.activations=['relu', 'relu'] and options.returnSequence=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual -4.7578125 should be close enough to expected -18.59375 by ULP distance: expected a number less than or equal to 6 but got 2020
[FAIL] gru float16 tensors steps=2 with options.bias, options.recurrentBias, options.direction='both' and options.returnSequence=true
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual -0.375 should be close enough to expected -0.28076171875 by ULP distance: expected a number less than or equal to 6 but got 386
[FAIL] gru float16 tensors steps=2 with all options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gru float16 actual -4.7578125 should be close enough to expected -18.59375 by ULP distance: expected a number less than or equal to 6 but got 2020
Harness: the test ran to completion.
